+++ NEWS.md
@@ -1,3 +1,5 @@
+# plume (development version)
+
 # plume 0.2.0
 
 ## New features