+++ NEWS.md
@@ -1,5 +1,7 @@
 # plume (development version)
 
+* Fixed `$get_contributions()` wrongly reordering authors when using CRediT roles and `by_author = TRUE` (#50).
+
 * `$get_contributions()` gains a new parameter `sep` that allows for finer control of how to separate contributors or roles (#49).
 
 * `$to_yaml()` now throws an error when invalid ORCID identifiers are found.